refactor : 문자열을 Lotto 객체로 바꿔주는 Converter 객체 생성
