+++ src/main/java/lotto/Application.java
@@ -3,6 +3,7 @@ package lotto;
 import lotto.io.Input;
 import lotto.io.Output;
 import lotto.io.View;
+import lotto.service.LottoService;
 
 public class Application {
     public static void main(String[] args) {
