+++ src/main/java/lotto/service/LottoService.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.service;
 
 import lotto.domain.Amount;
 import lotto.domain.Lotto;
